third_party/blink/web_tests/http/tests/websocket/reserved-opcodes-expected.txt

Tests whether WebSocket ignores frames with reserved opcode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Opcode 3: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 4: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 5: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 6: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 7: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 11: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 12: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 13: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 14: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 15: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
